+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.8.1
 - `,` between fields is optional like in GraphQL
+- Support `__typename` metadata
+
 # 0.8.0
 - `Moved EntityQueryLanguage.DataApi` namespace to `EntityQueryLanguage.GraphQL` to better communicate that its intent is to support
 - Add support GraphQL arguments in fields. See updated readme